add invalid id error

+++ backend/product.py
@@ -21,7 +21,7 @@ def new_product(name, price, description, category, deli_days, pic_link, db_name
     # assert db.check_interest_dim(category)
     # catagory is now calculated by query_analysis
     category = None
-    price = float(price)
+    price = int(price)
     deli_days = int(deli_days)
     if description == "" or description is None:
         description = name
